Remove spree user registrations controller file
